share metric alert spec fixtures
