+++ packages/services/workflows/src/lib/metric-alert-evaluator.spec.ts
@@ -1,4 +1,3 @@
-import type { Logger } from '@graphql-hive/logger';
 import { printWithValues, type SqlValue } from '@hive/clickhouse';
 import type { ClickHouseClient } from './clickhouse-client.js';
 import {
@@ -11,48 +10,10 @@ import {
   previousValueForRule,
   queryClickHouseWindows,
   resolutionFor,
-  type MetricAlertRuleRow,
 } from './metric-alert-evaluator.js';
-
-// Minimal logger that records warn() calls so we can assert defensive logging.
-function makeLogger() {
-  const warnings: unknown[][] = [];
-  const logger = {
-    warn: (...args: unknown[]) => warnings.push(args),
-    info: () => {},
-    error: () => {},
-    debug: () => {},
-    child: () => logger,
-  } as unknown as Logger;
-  return { logger, warnings };
-}
+import { makeLogger, makeRule } from './metric-alert-test-utils.js';
 
 const render = (conds: SqlValue[]) => conds.map(c => printWithValues(c));
-
-function makeRule(overrides: Partial<MetricAlertRuleRow>): MetricAlertRuleRow {
-  return {
-    id: 'r1',
-    organizationId: 'o1',
-    projectId: 'p1',
-    targetId: 't1',
-    name: 'rule',
-    type: 'TRAFFIC',
-    timeWindowMinutes: 60,
-    metric: null,
-    thresholdType: 'FIXED_VALUE',
-    thresholdValue: 100,
-    direction: 'ABOVE',
-    severity: 'WARNING',
-    state: 'NORMAL',
-    stateChangedAt: null,
-    lastEvaluatedAt: null,
-    confirmationMinutes: 0,
-    savedFilterId: null,
-    savedFilterFilters: null,
-    organizationPlanName: 'PRO',
-    ...overrides,
-  };
-}
 
 describe('groupRulesByQuery', () => {
   test('groups by target + window + savedFilterId', () => {
