+++ packages/services/workflows/src/lib/metric-alert-notifier.spec.ts
@@ -1,7 +1,6 @@
-import { describe, expect, it, vi } from 'vitest';
-import type { Logger } from '@graphql-hive/logger';
 import type { PostgresDatabasePool } from '@hive/postgres';
 import type { AlertChannelRow, NotificationEvent } from './metric-alert-notifier.js';
+import { makeLogger, makeRule } from './metric-alert-test-utils.js';
 
 const postMessage = vi.fn(async (_args: { attachments: Array<{ color: string }> }) => ({
   ok: true,
@@ -14,14 +13,6 @@ vi.mock('@slack/web-api', () => ({
 }));
 
 const { sendSlackNotification } = await import('./metric-alert-notifier.js');
-
-const logger = {
-  warn: () => {},
-  info: () => {},
-  error: () => {},
-  debug: () => {},
-  child: () => logger,
-} as unknown as Logger;
 
 const pg = {
   maybeOneFirst: async () => 'xoxb-test-token',
@@ -38,16 +29,7 @@ const channel: AlertChannelRow = {
 function makeEvent(overrides: Partial<NotificationEvent> = {}): NotificationEvent {
   return {
     state: 'firing',
-    rule: {
-      organizationId: 'org-1',
-      name: 'Console - Error Rate > 2% over 15min',
-      type: 'ERROR_RATE',
-      metric: null,
-      severity: 'WARNING',
-      thresholdType: 'FIXED_VALUE',
-      thresholdValue: 2,
-      direction: 'ABOVE',
-    },
+    rule: makeRule({ type: 'ERROR_RATE', thresholdValue: 2, severity: 'WARNING' }),
     currentValue: 2.15,
     previousValue: null,
     organizationSlug: 'the-guild',
@@ -57,19 +39,19 @@ function makeEvent(overrides: Partial<NotificationEvent> = {}): NotificationEven
   };
 }
 
-async function colorFor(event: NotificationEvent): Promise<unknown> {
+async function colorFor(event: NotificationEvent): Promise<string> {
   postMessage.mockClear();
-  await sendSlackNotification({ channel, event, pg, logger });
+  await sendSlackNotification({ channel, event, pg, logger: makeLogger().logger });
   return postMessage.mock.calls[0][0].attachments[0].color;
 }
 
 describe('sendSlackNotification', () => {
-  it('uses the severity hex for the firing state', async () => {
+  test('uses the severity hex for the firing state', async () => {
     await expect(colorFor(makeEvent())).resolves.toBe('#c5870d');
   });
 
   // Slack renders `good`/`warning`/`danger` as a grey bar, so resolved must be a hex.
-  it('uses a green hex for the resolved state', async () => {
+  test('uses a green hex for the resolved state', async () => {
     await expect(colorFor(makeEvent({ state: 'resolved', currentValue: 1.64 }))).resolves.toBe(
       '#2ECC71',
     );
